# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Project": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"Active",
							"Inactive"
						]
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Project.Exclude_keyofProject.id__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"Active",
							"Inactive"
						]
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"description",
					"startDate",
					"endDate"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Project.id_": {
				"$ref": "#/components/schemas/Pick_Project.Exclude_keyofProject.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateProject": {
				"$ref": "#/components/schemas/Omit_Project.id_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ing-hrm-wiki",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/projects/{projectId}": {
			"get": {
				"operationId": "GetProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Project"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/projects": {
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProject"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}